backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.134.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
